Extended FilterItems to support non-string fields
Added two new method overloads for ItemService.FilterItems for filtering
on non-string fields.
The overloads differ by the type of the 'filters' parameter,
respectively Object and IDictionary<string, object>.

The Object overload is intended for passing Anonymous Types describing
the filter.
The method body contain examples of how to construct the types for
different datatypes. Ideally someone will move this to a wiki!